Refactor updateDisplay to fix erroneous clearing
Running the total function sets subDisplay to a space.  updateDisplay
checks subDisplay to determine if total was the last function ran and
if so, clears the display before inputting the new number.  An
unintended side effect of this was that if total was the last function
ran, only a single digit could be input as each number would cause
updateDisplay to clear the display before inputting the new number.
To address this, after the display is cleared, subDisplay is returned
to an empty string so that on further calls of updateDisplay the display
will not be cleared before the new number is added to the display.